Move Types and function usage
