+++ src/tracer/src/process_command.rs
@@ -14,10 +14,10 @@ use crate::process_identification::constants::{
     DEFAULT_DAEMON_PORT, PID_FILE, STDERR_FILE, STDOUT_FILE, WORKING_DIR,
 };
 use crate::process_identification::debug_log::Logger;
-use crate::utils::analytics::emit_analytic_event;
+use crate::utils::analytics::types::AnalyticsEventType;
 use crate::utils::file_system::ensure_file_can_be_created;
 use crate::utils::system_info::check_sudo_privileges;
-use crate::utils::{Sentry, Version};
+use crate::utils::{analytics, Sentry, Version};
 use anyhow::{Context, Result};
 use clap::Parser;
 use daemonize::{Daemonize, Outcome};
@@ -223,15 +223,15 @@ pub fn process_cli() -> Result<()> {
 
                     // Wait a moment for daemon to start, then show info
                     tokio::runtime::Runtime::new()?.block_on(async {
-                        tokio::spawn(emit_analytic_event(
+                        analytics::spawn_event(
                             args.user_id.clone(),
-                            crate::process_identification::types::analytics::AnalyticsEventType::DaemonStartAttempted,
+                            AnalyticsEventType::DaemonStartAttempted,
                             None,
-                        ));
+                        );
                         let _ = print_install_readiness();
                         wait(&api_client).await?;
 
-                        print_config_info(&api_client, &config,false).await
+                        print_config_info(&api_client, &config, false).await
                     })?;
 
                     return Ok(());
