+++ src/tracer/src/process_identification/types/mod.rs
@@ -1,4 +1,3 @@
-pub mod analytics;
 pub mod cli;
 pub mod current_run;
 pub mod event;
